+++ Banking Platform/src/Users/Basic.java
@@ -12,11 +12,11 @@ public interface Basic {
      * @param currency
      * @return
      */
-    Account createNewAccount(User[] beneficiaries, String accountNumber, float availableBalance, String currency);
+    Account createNewAccount(User[] beneficiaries, String accountNumber, double availableBalance, String currency);
     Account createNewAccount(User[] beneficiaries, String accountNumber, String currency);
     void closeAccount(User user);
     void deleteAccount(String accountNumber);
-    void transferToAccount(User user, String accountFrom, String accountTo, float amount);
+    void transferToAccount(User user, String accountFrom, String accountTo, double amount);
     void viewBalance(User user);
     void viewBalance(String account);
     void viewTransactions(User user);
